% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityfn.R
\name{densityfn}
\alias{densityfn}
\title{wrapper to evaluate a density function given by a character string as an argument}
\usage{
densityfn(x, family, ...)
}
\arguments{
\item{x}{the element of the sample space at which to evaluate the density}

\item{family}{the name of the density function, character string}

\item{...}{other named arguments to feed into the density function}
}
\value{
the value of the density
}
\description{
wrapper to evaluate a density function given by a character string as an argument
}
\keyword{internal}
