% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.add.R
\name{form.add}
\alias{form.add}
\title{adds together two formula}
\usage{
form.add(formula1, formula2)
}
\arguments{
\item{formula1}{a formula}

\item{formula2}{a formula}
}
\value{
another formula
}
\description{
adds together two formula
}
\keyword{internal}
