% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mreg_package.R
\docType{package}
\name{mreg_package}
\alias{mreg_package}
\title{mreg: Implements the techniques of exact likelihood when the discrete outcome can be missing in a regression mode}
\description{
Implements the methods described in Bond S, Farewell V, 2006, Exact Likelihood Estimation for a
Negative Binomial Regression Model with Missing Outcomes, Biometrics,
Submitted. The main function is \code{\link{mreg}}.
}
\references{
Bond S, Farewell V, 2006, Exact Likelihood Estimation for a
Negative Binomial Regression Model with Missing Outcomes, Biometrics
}
