% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mreg.R
\name{print.mreg}
\alias{print.mreg}
\title{Prints the coefficients from an mreg object}
\usage{
\method{print}{mreg}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{an mreg object}

\item{digits}{number of digits to print decimals to.}

\item{...}{other arguments, not currently used but required for generic methods}
}
\value{
It prints the coefficients from an
\code{\link{mreg}} regression model.
}
\description{
Prints the coefficients from an mreg object
}
\seealso{
\code{\link{mreg}} \code{\link{summary.mreg}}
}
\keyword{print}
