% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randfn.R
\name{randfn}
\alias{randfn}
\title{wrapper to generate random samples from a density function given by a character string as an argument}
\usage{
randfn(n, family, ...)
}
\arguments{
\item{n}{the number of samples to generate}

\item{family}{the name of the density function, character string}

\item{...}{other named arguments to feed into the density function}
}
\value{
the vector of samples
}
\description{
wrapper to generate random samples from a density function given by a character string as an argument
}
\keyword{internal}
