% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unity.R
\name{unity}
\alias{unity}
\title{The default value of 'modify' function in 'mreg'}
\usage{
unity(x, y, mod.Z)
}
\arguments{
\item{x}{is a vector of possible values the previous value of the
response could take.}

\item{y}{is an vector of the coefficients. Its length is specified in
the \code{\link{mreg}} function by the argument \code{modify.p}.}

\item{mod.Z}{vector of observed covariates that may interact with the
unobserved preceding outcome corresponding to the observation. It is
taken from the \code{\link{model.frame}} produced by the \code{mod.formula}
argument in \code{\link{mreg}}.}
}
\value{
A vector of zeroes the same length is the argument \code{x}.
}
\description{
If the location term in a regression model does not depend on any
previously observed values of the outcome in a longitudinal data set,
then we obtain simplification in our estimation procedure when the
outcomes can be missing. Using the default value of \code{unity} for the
argument \code{modify} in the \code{\link{mreg}} function does this.
}
\details{
This function  is the default value for the argument \code{modify}  for
\code{\link{mreg}}. It does nothing to the linear predictor term.
    For this function \code{unity} there are no such covariates.
    A default value for \code{mod.formula} is \code{~1}.
}
\keyword{programming}
\keyword{regression}
